+++ src/Formie.php
@@ -110,7 +110,7 @@ class Formie extends Plugin
 
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '4.0.9';
+    public string $schemaVersion = '4.0.10';
     public string $minVersionRequired = '2.1.5';